change credit response to integer

+++ model/src/journal.rs
@@ -1,7 +1,7 @@
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Transaction
 {
-    pub value: f64,
+    pub value: i32,
     pub description: String,
     pub timestamp: i64,
 }
